+++ src/config/degrees.ts
@@ -1,6 +1,6 @@
 export enum Degrees {
-  UNDERGRADUATE = 'Undergraduate',
   HIGHSCHOOL = 'High School',
+  UNDERGRADUATE = 'Undergraduate',
   GRADUATE = 'Graduate',
   OTHER = 'Other',
 }